Refactored code snipped to the right way
I made a mistake at the beginning, because I didn't know how the end product is gonna look, where I had 2 instances of the search area with the only difference being 1 class name. Changed it to only 1 instance with the 2 class names switching based on state.
Also added a click away listener to the sub box.